+++ src/screens/SettingsScreen/SettingsOption/SettingsOption.types.ts
@@ -1,0 +1,8 @@
+import { GestureResponderEvent } from 'react-native';
+import { IconType } from '@styles';
+
+export interface SettingsOptionProps {
+	label: string;
+	icon: IconType;
+	onPress: (event: GestureResponderEvent) => void;
+}
